# [WIP][AccessEnforcementOpts] Add mergeAccesses analysis + optimization