… lib/SILOptimizer/Utils/LoopUtils.cpp
… test/SILOptimizer/merge_exclusivity.swift