@shajrawi 2018-07-28
@swift-ci Please clean test OS X