@shajrawi 2018-07-30
Comparing exclusivity enabled on master v.s. this PR:

TEST | OLD | NEW | DELTA | SPEEDUP
-- | -- | -- | -- | --
RangeIterationSigned | 687 | 92 | -86.6% | **7.47x**
DropFirstCountableRangeLazy | 28 | 14 | -50.0% | **2.00x**
PolymorphicCalls | 20 | 14 | -30.0% | **1.43x**
Sim2DArray | 330 | 246 | -25.5% | **1.34x**
ExclusivityGlobal | 4 | 3 | -25.0% | **1.33x**
Chars | 605 | 493 | -18.5% | **1.23x**