@jrose-apple 2018-07-30
Can you go into a little more detail about why this is safe? What if `%14` modifies `%1` through another access path?