@shajrawi 2018-07-31
See `FunctionAccessedStorage` analysis (`AccessedStorageAnalysis.h` line 127) : the output of that analysis is checking if there are any unidentified accesses in that function and its callees (`hasUnidentifiedAccess`) and, if we are able to analyze the entire chain and all the accessed storage locations, we have a summary of all the storage locations accessed.
Given `%1`, we have to call the conservative `mayConflictWith` method on `%14` which will tell us if this is safe or not.