@shajrawi 2018-08-01
replaced by https://github.com/apple/swift/pull/18438 - closing this PR